[refac] Refactor JWT package structure for improved organization
