+++ edukit-core/src/main/java/com/edukit/core/auth/service/jwt/JwtGenerator.java
@@ -1,8 +1,8 @@
-package com.edukit.core.auth.jwt;
+package com.edukit.core.auth.service.jwt;
 
-import com.edukit.core.auth.jwt.provider.JwtKeyProvider;
-import com.edukit.core.auth.jwt.setting.JwtProperties;
-import com.edukit.core.auth.jwt.type.TokenType;
+import com.edukit.core.auth.service.jwt.provider.JwtKeyProvider;
+import com.edukit.core.auth.service.jwt.setting.JwtProperties;
+import com.edukit.core.auth.service.jwt.type.TokenType;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
