+++ edukit-core/src/main/java/com/edukit/core/auth/service/jwt/setting/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.edukit.core.auth.jwt.setting;
+package com.edukit.core.auth.service.jwt.setting;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
